:recycle:  Refactor
