+++ src/app/features/percentage-calculator/percentage-calculator.component.scss
