+++ src/app/features/personal-records/new-pr/new-pr.component.scss
