+++ src/app/features/personal-records/new-pr/new-pr.component.ts
@@ -1,10 +1,10 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
-import { ExerciseEnum } from '../../shared/enums/ExerciseEnum';
-import { LSKeysEnum } from '../../shared/enums/LSKeysEnum';
-import { LocalStorageService } from '../../service/local-storage.service';
-import { SharedService } from '../../service/shared.service';
-import { ExerciseLabelPipe } from '../../shared/pipes/exercise-label.pipe';
+import { ExerciseEnum } from '../../../shared/enums/ExerciseEnum';
+import { LSKeysEnum } from '../../../shared/enums/LSKeysEnum';
+import { LocalStorageService } from '../../../service/local-storage.service';
+import { SharedService } from '../../../service/shared.service';
+import { ExerciseLabelPipe } from '../../../shared/pipes/exercise-label.pipe';
 import { PersonalRecord } from '../personal-record.interface';
 import { FormsModule } from '@angular/forms';
