+++ src/app/features/personal-records/personal-record.interface.ts
@@ -1,4 +1,4 @@
-import { ExerciseEnum } from '../shared/enums/ExerciseEnum';
+import { ExerciseEnum } from '../../shared/enums/ExerciseEnum';
 
 export interface PersonalRecord {
   recordName: string;
